feat: generic popup component
