+++ packages/blocks/src/database-block/table/components/toolbar/toolbar.ts
@@ -1,20 +1,23 @@
-import './toolbar-action-popup.js';
 import '../../../common/filter/filter-group.js';
 
 import {
+  CopyIcon,
   DatabaseExpandWide,
   DatabaseSearchClose,
   DatabaseSearchIcon,
+  DeleteIcon,
   MoreHorizontalIcon,
   PlusIcon,
 } from '@blocksuite/global/config';
 import type { BlockSuiteRoot } from '@blocksuite/lit';
 import { ShadowlessElement, WithDisposable } from '@blocksuite/lit';
 import { DisposableGroup } from '@blocksuite/store';
-import { computePosition } from '@floating-ui/dom';
 import { css, html } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
 
+import { copyBlocks } from '../../../../__internal__/clipboard/index.js';
+import { popFilterableSimpleMenu } from '../../../../components/menu/menu.js';
+import { toast } from '../../../../components/toast.js';
 import { stopPropagation } from '../../../../page-block/edgeless/utils.js';
 import type { FilterGroup } from '../../../common/ast.js';
 import { firstFilterByRef } from '../../../common/ast.js';
@@ -30,7 +33,6 @@ import type { TableViewManager } from '../../table-view-manager.js';
 import { SearchState } from '../../types.js';
 import { showDatabaseTableViewModal } from '../modal/index.js';
 import { initAddNewRecordHandlers } from './index.js';
-import { ToolbarActionPopup } from './toolbar-action-popup.js';
 
 const styles = css`
   .affine-database-toolbar {
@@ -210,7 +212,6 @@ export class DatabaseToolbar extends WithDisposable(ShadowlessElement) {
   @query('.new-record')
   private _newRecord!: HTMLDivElement;
 
-  private _toolbarAction!: ToolbarActionPopup | undefined;
   private _recordAddDisposables = new DisposableGroup();
 
   private get readonly() {
@@ -308,38 +309,31 @@ export class DatabaseToolbar extends WithDisposable(ShadowlessElement) {
 
   private _onShowAction = () => {
     if (this.readonly) return;
-
-    if (this._toolbarAction) {
-      this._closeToolbarAction();
-      return;
-    }
-    this.setSearchState(SearchState.Action);
-    const toolbarAction = new ToolbarActionPopup();
-    toolbarAction.targetModel = this.targetModel;
-    toolbarAction.close = this._closeToolbarAction;
-    this._toolbarAction = toolbarAction;
-    this._moreActionContainer.appendChild(this._toolbarAction);
-    computePosition(this._moreActionContainer, this._toolbarAction, {
-      placement: 'bottom',
-    }).then(({ x, y }) => {
-      Object.assign(toolbarAction.style, {
-        left: `${x}px`,
-        top: `${y}px`,
-      });
-    });
-
-    onClickOutside(
-      this._moreActionContainer,
-      () => {
-        this._closeToolbarAction();
+    popFilterableSimpleMenu(this._moreActionContainer, [
+      {
+        type: 'action',
+        name: 'Copy',
+        icon: CopyIcon,
+        select: () => {
+          copyBlocks({
+            type: 'Block',
+            models: [this.targetModel],
+            startOffset: 0,
+            endOffset: 0,
+          });
+          toast('Copied Database to clipboard');
+        },
       },
-      'mousedown'
-    );
-  };
-
-  private _closeToolbarAction = () => {
-    this._toolbarAction?.remove();
-    this._toolbarAction = undefined;
+      {
+        type: 'action',
+        name: 'Delete database',
+        icon: DeleteIcon,
+        select: () => {
+          const models = [this.targetModel, ...this.targetModel.children];
+          models.forEach(model => this.targetModel.page.deleteBlock(model));
+        },
+      },
+    ]);
   };
 
   private _resetSearchStatus = () => {
@@ -469,15 +463,15 @@ export class DatabaseToolbar extends WithDisposable(ShadowlessElement) {
       </div>
       ${this.modalMode
         ? null
-        : html`<div
+        : html` <div
             class="affine-database-toolbar-item expand"
-            @click=${this._onShowModalView}
+            @click="${this._onShowModalView}"
           >
             ${DatabaseExpandWide}
           </div>`}
       ${this.readonly
         ? null
-        : html`<div
+        : html` <div
               class="affine-database-toolbar-item more-action ${isActiveMoreAction
                 ? 'active'
                 : ''}"
